requirements.txt
fastf1
pandas
matplotlib
numpy
arcade
pyglet
pyside6
questionary
rich
pyserial==3.5